% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCohorts.R
\name{CompareCohorts}
\alias{CompareCohorts}
\title{CompareCohorts}
\usage{
CompareCohorts(
  connectionDetailsVocab,
  cohorts,
  Concepts_in_cohortSetOldCht,
  Concepts_in_cohortSetNewCht,
  newVocabSchema,
  oldVocabSchema,
  resultSchema,
  excludedNodes = 0,
  includedSourceVocabs = 0
)
}
\arguments{
\item{Concepts_in_cohortSetOldCht}{old cohort definition run on an old vocabulary -
dataframe which stores cohorts and concept set definitions in a tabular format,
it should have the following columns:
"ConceptID","isExcluded","includeDescendants","conceptsetId","conceptsetName","cohortId"}

\item{Concepts_in_cohortSetNewCht}{updated cohort definition run on a new vocabulary -
dataframe which stores cohorts and concept set definitions in a tabular format,
it should have the following columns:
"ConceptID","isExcluded","includeDescendants","conceptsetId","conceptsetName","cohortId"}

\item{newVocabSchema}{schema containing a new vocabulary version}

\item{oldVocabSchema}{schema containing an older vocabulary version}

\item{resultSchema}{schema containing Achilles results}

\item{excludedNodes}{text string with excluded nodes, for example: "9201, 9202, 9203"; 0 by default}

\item{includedSourceVocabs}{text string with included source vocabularies, for example: "'ICD10CM', 'ICD9CM', 'HCPCS'"; 0 by default, which is treated as ALL vocabularies}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}
}
\description{
This function resolves concept sets of two cohorts run on different vocabulary versions and writes the result to the Excel file
it uses an input of \code{getNodeConcepts()} funcion,
it detects
for the updated cohort definition run on a new vocabulary
1) non-standard concepts used in concept set expression;
for the old cohort definition run on an old vocabulary vs updated cohort definition run on a new vocabulary:
2) added or excluded source concepts due to changed mapping to standard concepts
3) domain changes of included standard concepts
The result is written to an excel file with the tab for each check
}
\details{
This function resolves concept sets of two cohorts run on different vocabulary versions and writes the result to the Excel file
}
\examples{
\dontrun{
                             CompareCohorts(connectionDetailsVocab = connectionDetailsVocab,
                                           cohorts = cohorts,
                                           Concepts_in_cohortSetOldCht = Concepts_in_cohortSetOldCht,
                                           Concepts_in_cohortSetNewCht = Concepts_in_cohortSetNewCht,
                                           newVocabSchema = newVocabSchema,
                                           oldVocabSchema = oldVocabSchema,
                                           excludedNodes = excludedVisitNodes,
                                           resultSchema = resultSchema)

}
}
