% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeConcepts.R
\name{getNodeConcepts}
\alias{getNodeConcepts}
\title{This function extracts the cohort_id - conceptSet - NodeConcept(the concept used in ConceptSetDefinition) - IsIncluded - includeDescendants}
\usage{
getNodeConcepts(cohorts = cohorts, baseUrl)
}
\arguments{
\item{cohorts}{list of cohorts to be evaluated}

\item{baseUrl}{the BaseUrl of your Atlas instance}
}
\description{
This function extracts the cohort_id - conceptSet - NodeConcept(the concept used in ConceptSetDefinition) - IsIncluded - includeDescendants
}
