% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceCodesCount.R
\name{sourceCodesCount}
\alias{sourceCodesCount}
\title{Run the query to get source concepts counts across all CDM tables}
\usage{
sourceCodesCount(resSchema, connectionDetails)
}
\arguments{
\item{resSchema}{The schema with the Achilles Results}

\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}
}
\description{
Run the query to get source concepts counts across all CDM tables
}
