% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultToExcel.R
\name{resultToExcel}
\alias{resultToExcel}
\title{This function resolves concept sets in a SQl database, it needs an input of
cohort_id - conceptSet - NodeConcept(the concept used in ConceptSetDefinition) - IsIncluded - includeDescendants,
it then detects non-standard concepts used as part of concept set definition
also detects the following changes in included concepts due to a vocabulary version change:
1) added or excluded source concepts with changed mapping to standard concepts
2) hierarchy changes are shown as peak concepts added or removed. the peak concept is the concept above which the hierarchy is altered
3) domain changes of included standard concepts
The result is stored as an excel file with the tab for each check}
\usage{
resultToExcel(
  connectionDetails,
  Concepts_in_cohortSet,
  workSchema,
  newVocabSchema,
  oldVocabSchema,
  resultSchema,
  excl_node,
  source_concept_rules
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{Concepts_in_cohortSet}{dataframe which stores cohorts and concept set definitions in a tabular format,
it should have the following columns:
"ConceptID","isExcluded","includeDescendants","conceptsetId","conceptsetName","cohortId"}

\item{workSchema}{schema with a write access where tables are stored when executing the query}

\item{newVocabSchema}{schema containing a new vocabulary version}

\item{oldVocabSchema}{schema containing an older vocabulary version}

\item{resultSchema}{schema containing Achilles results}

\item{excl_node}{dataframe with nodes excluded from analysis !! turn it into vector?}

\item{source_concept_rules}{dataframe containing domain_id and vocabuary_id of source concepts excluded or included only in the output}
}
\description{
This function resolves concept sets in a SQl database, it needs an input of
cohort_id - conceptSet - NodeConcept(the concept used in ConceptSetDefinition) - IsIncluded - includeDescendants,
it then detects non-standard concepts used as part of concept set definition
also detects the following changes in included concepts due to a vocabulary version change:
1) added or excluded source concepts with changed mapping to standard concepts
2) hierarchy changes are shown as peak concepts added or removed. the peak concept is the concept above which the hierarchy is altered
3) domain changes of included standard concepts
The result is stored as an excel file with the tab for each check
}
