% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultToExcel.R
\name{resultToExcel}
\alias{resultToExcel}
\title{This function resolves concept sets in a SQL database and writes the result to the Excel file}
\usage{
resultToExcel(
  connectionDetails,
  Concepts_in_cohortSet,
  workSchema,
  newVocabSchema,
  oldVocabSchema,
  resultSchema,
  excludedNodes = 0
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{Concepts_in_cohortSet}{dataframe which stores cohorts and concept set definitions in a tabular format,
it should have the following columns:
"ConceptID","isExcluded","includeDescendants","conceptsetId","conceptsetName","cohortId"}

\item{workSchema}{schema with a write access where tables are stored when executing the query}

\item{newVocabSchema}{schema containing a new vocabulary version}

\item{oldVocabSchema}{schema containing an older vocabulary version}

\item{resultSchema}{schema containing Achilles results}

\item{excludedNodes}{text string with excluded nodes, for example: "9201, 9202, 9203"; 0 by default}
}
\description{
This function resolves concept sets in a SQL database
it uses an input of #' getNodeConcepts() funcion,
it detects non-standard concepts used in concept set expression;
also detects the changes in included concepts due to a vocabulary version change:
1) added or excluded source concepts due to changed mapping to standard concepts
2) added or excluded standard concepts due to hierarchy changes, only the "peak concepts" are shown
 "peak concept" is a concept above which the hierarchy is altered
3) domain changes of included standard concepts
The result is written to an excel file with the tab for each check
}
\examples{
\dontrun{
 resultToExcel(connectionDetails = YourconnectionDetails,
 Concepts_in_cohortSet = Concepts_in_cohortSet, # is returned by getNodeConcepts function
 workSchema = "workSchema", #schema where user is allowed to create tables
 newVocabSchema = "omopVocab_v1", #schema containing newer vocabulary version
 oldVocabSchema = "omopVocab_v0", #schema containing older vocabulary version
 resultSchema = "achillesresults") #schema with achillesresults
}
}
